refactor: Restructures layouts
