+++ components/pages/articles/base/index.tsx
@@ -4,7 +4,7 @@ import IArticle from "types/article";
 import dynamic from "next/dynamic";
 import withNavbarLayout from "lib/with-navbar-layout";
 
-const Articles = dynamic(import("components/articles"));
+const Articles = dynamic(import("components/layouts/articles"));
 
 interface Props {
   articles: IArticle[];
