+++ components/pages/publications/base/index.tsx
@@ -4,7 +4,7 @@ import IPublication from "types/publication";
 import dynamic from "next/dynamic";
 import withNavbarLayout from "lib/with-navbar-layout";
 
-const Publications = dynamic(import("components/publications"));
+const Publications = dynamic(import("components/layouts/publications"));
 
 interface Props {
   publications: IPublication[];
